Solution correctly stored and visualized.

+++ common.h
@@ -38,7 +38,7 @@ typedef struct{
 
     char *filename; //TODO free memory after allocation
 
-    double **solution;
+    int **solution;
     int solution_size;
 
 } Tsp_prob;
